+++ .gitignore
@@ -1,0 +1,3 @@
+*.sw[op]
+*.pyc
+__pycache__
