+++ bull/bull.py
@@ -9,66 +9,19 @@ import sys
 import uuid
 
 from jinja2 import Environment, PackageLoader
-from flask import (Flask, send_from_directory, abort, redirect, request,
+from flask import (Blueprint, send_from_directory, abort, redirect, request,
                    render_template)
 from flask.ext.sqlalchemy import SQLAlchemy
 from flask.ext.mail import Mail, Message
 import stripe
 
-from . import app
+from .models import Product, Purchase, db
 
-stripe.api_key = app.config['STRIPE_SECRET_KEY']
-
-db = SQLAlchemy(app)
-mail = Mail(app)
+bull = Blueprint('bull', __name__)
 env = Environment(loader=PackageLoader('bull', 'templates'))
+mail = Mail()
 
-class Product(db.Model):
-    """A digital product for sale on our site.
-
-    :param int id: Unique id for this product
-    :param str name: Human-readable name of this product
-    :param str file_name: Path to file this digital product represents
-    :param str version: Optional version to track updates to products
-    :param bool is_active: Used to denote if a product should be considered
-                          for-sale
-    :param float price: Price of product
-    """
-    __tablename__ = 'product'
-    id = db.Column(db.Integer, primary_key=True, autoincrement=True)
-    name = db.Column(db.String)
-    file_name = db.Column(db.String)
-    version = db.Column(db.String, default=None, nullable=True)
-    is_active = db.Column(db.Boolean, default=True, nullable=True)
-    price = db.Column(db.Float)
-
-    def __str__(self):
-        """Return the string representation of a product."""
-        return '{} (v{})'.format(self.name, self.version)
-
-class Purchase(db.Model):
-    """Contains information about the sale of a product.
-
-    :param str uuid: Unique ID (and URL) generated for the customer unique to
-                     this purchase
-    :param str email: Customer's email address
-    :param int product_id: ID of the product associated with this sale
-    :param :class:`SQLAlchemy.relationship` product: The associated product
-    :param downloads_left int: Number of downloads remaining using this URL
-    """
-    __tablename__ = 'purchase'
-    uuid = db.Column(db.String, primary_key=True)
-    email = db.Column(db.String)
-    product_id = db.Column(db.Integer, db.ForeignKey('product.id'))
-    product = db.relationship(Product)
-    downloads_left = db.Column(db.Integer, default=5)
-    sold_at = db.Column(db.DateTime, default=datetime.datetime.now)
-
-    def __str__(self):
-        """Return the string representation of the purchase."""
-        return '{} bought by {}'.format(self.product.name, self.email)
-
-@app.route('/<uuid>')
+@bull.route('/<uuid>')
 def download_file(uuid):
     """Serve the file associated with the purchase whose ID is *uuid*.
 
@@ -85,8 +38,9 @@ def download_file(uuid):
                 filename=purchase.product.file_name, as_attachment=True)
     else:
         abort(404)
-    
-@app.route('/buy', methods=['POST'])
+
+
+@bull.route('/buy', methods=['POST'])
 def buy():
     """Facilitate the purchase of a product."""
 
@@ -126,15 +80,13 @@ def buy():
 
     return render_template('success.html', url=purchase.uuid)
 
-@app.route('/test/<product_id>')
+@bull.route('/test/<product_id>')
 def test(product_id):
     """Return a test page for live testing the "purchase" button."""
     test_product = Product.query.get(product_id)
     return render_template(
             'test.html', 
-            test_product=test_product, 
-            stripe_key=app.config['STRIPE_PUBLIC_KEY'],
-            site_name=app.config['SITE_NAME'])
+            test_product=test_product)
 
 if __name__ == '__main__':
     sys.exit(app.run(debug=True))
